Add day_chart type to bar chart race endpoint
Restructure BarChartRace from a single class into a module with two
classes: CumulativeFrames (existing animated behavior) and DayChart
(single-frame snapshot of top songs for the given period). A factory
method BarChartRace.for dispatches based on the type parameter.

Closes #341 on playlists-interface
