+++ spec/services/bar_chart_race/day_chart_spec.rb
@@ -1,0 +1,127 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe BarChartRace::DayChart do
+  let(:radio_station) { create(:radio_station) }
+  let(:song_a) { create(:song, title: 'Song A') }
+  let(:song_b) { create(:song, title: 'Song B') }
+  let(:song_c) { create(:song, title: 'Song C') }
+
+  describe '#frames' do
+    context 'with period parameter' do
+      let(:params) { { period: 'day' } }
+      let(:chart) { described_class.new(radio_station:, params:) }
+
+      before do
+        5.times { |i| create(:air_play, song: song_a, radio_station:, broadcasted_at: 2.hours.ago + i.minutes) }
+        3.times { |i| create(:air_play, song: song_b, radio_station:, broadcasted_at: 2.hours.ago + i.minutes) }
+        create(:air_play, song: song_c, radio_station:, broadcasted_at: 2.hours.ago)
+      end
+
+      it 'returns a single frame' do
+        expect(chart.frames.size).to eq(1)
+      end
+
+      it 'ranks songs by play count', :aggregate_failures do
+        frame = chart.frames.first
+        entries = frame[:entries]
+
+        expect(entries.first[:song][:title]).to eq('Song A')
+        expect(entries.first[:count]).to eq(5)
+        expect(entries.first[:position]).to eq(1)
+        expect(entries.second[:song][:title]).to eq('Song B')
+        expect(entries.second[:count]).to eq(3)
+        expect(entries.second[:position]).to eq(2)
+      end
+
+      it 'includes song details in entries', :aggregate_failures do
+        entry = chart.frames.first[:entries].first
+
+        expect(entry[:song]).to include(:id, :title, :spotify_artwork_url, :artists)
+      end
+
+      it 'includes date in the frame' do
+        expect(chart.frames.first[:date]).to match(/\A\d{4}-\d{2}-\d{2}\z/)
+      end
+    end
+
+    context 'with week period' do
+      let(:params) { { period: 'week' } }
+      let(:chart) { described_class.new(radio_station:, params:) }
+
+      before do
+        3.times { |i| create(:air_play, song: song_a, radio_station:, broadcasted_at: 3.days.ago.midday + i.minutes) }
+        2.times { |i| create(:air_play, song: song_a, radio_station:, broadcasted_at: 1.day.ago.midday + i.minutes) }
+      end
+
+      it 'sums all plays within the period into a single frame' do
+        frame = chart.frames.first
+
+        expect(frame[:entries].first[:count]).to eq(5)
+      end
+    end
+
+    context 'with more than 10 songs' do
+      let(:params) { { period: 'day' } }
+      let(:chart) { described_class.new(radio_station:, params:) }
+
+      before do
+        12.times do |i|
+          song = create(:song, title: "Song #{i}")
+          create(:air_play, song:, radio_station:, broadcasted_at: 2.hours.ago + i.minutes)
+        end
+      end
+
+      it 'limits to top 10' do
+        expect(chart.frames.first[:entries].size).to eq(10)
+      end
+    end
+
+    context 'with no air plays' do
+      let(:params) { { period: 'day' } }
+      let(:chart) { described_class.new(radio_station:, params:) }
+
+      it 'returns empty array' do
+        expect(chart.frames).to be_empty
+      end
+    end
+
+    context 'with custom date range' do
+      let(:params) { { start_time: 4.days.ago.strftime('%Y-%m-%dT%H:%M'), end_time: 1.day.ago.strftime('%Y-%m-%dT%H:%M') } }
+      let(:chart) { described_class.new(radio_station:, params:) }
+
+      before do
+        3.times { |i| create(:air_play, song: song_a, radio_station:, broadcasted_at: 2.days.ago.midday + i.minutes) }
+        2.times { |i| create(:air_play, song: song_a, radio_station:, broadcasted_at: 10.days.ago.midday + i.minutes) }
+      end
+
+      it 'only includes plays within the date range' do
+        frame = chart.frames.first
+
+        expect(frame[:entries].first[:count]).to eq(3)
+      end
+    end
+  end
+
+  describe '#meta' do
+    let(:params) { { period: 'day' } }
+    let(:chart) { described_class.new(radio_station:, params:) }
+
+    it 'returns type, period, and time range', :aggregate_failures do
+      meta = chart.meta
+
+      expect(meta[:type]).to eq('day_chart')
+      expect(meta[:period]).to eq('day')
+      expect(meta[:start_time]).to be_present
+      expect(meta[:end_time]).to be_present
+    end
+
+    it 'returns ISO8601 formatted times', :aggregate_failures do
+      meta = chart.meta
+
+      expect { Time.iso8601(meta[:start_time]) }.not_to raise_error
+      expect { Time.iso8601(meta[:end_time]) }.not_to raise_error
+    end
+  end
+end
